+++ CHANGELOG.md
@@ -8,3 +8,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Built initial Block interface structure
 - Added Block Category system
 - Implemented initial Area interface structure
+- Added Collection interface system
+- Split block lists into BlockGroup and Descriptor interfaces